refactor(i18n): refactor i18n code
